python_web_io/pyproject.toml
[tool.poetry]
name = "python-web-io"
version = "0.1.11"
description = "Generate a webpage as a GUI for a Python script, and serve from anywhere."
authors = ["Zachary Smith"]
packages = [{include = "python_web_io"}]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Cutwell/python-web-io"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
toml = "^0.10.2"
flask-session2 = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.2.2"
langchain = "^0.0.153"
openai = "^0.27.5"
pandas = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
python_web_io = "python_web_io.main:start"